Lab Project 3 (Lack of Error Handling Example 1)
- Description:
The flaw in the code is the lack of error handling that can cause difficulties to debug any problem encountered, it can cause unexpected crashes or failure in the code

- Method:
To refactor this flaw, I knew that the code mainly depends on the API request, if something goes wrong with the API request than the code might not run properly

- Conclusion:
Refactoring the code in this way enables the developers of the app to do debugging much more easier, by adding an error handling with notifications alert and redirecting to another page, By redirecting to a dedicated error page you can isolate the error and prevent it from affecting the current page or user flow. This helps in containing the error and maintaining the overall stability of the application.