Lab Project 3 (Long Method Example 2)
- Description:
The flaw in the code is The `getDish()` function above can be considered quite long and hard to read. it can be refactored to be a bit more readable and maintainable

- Method:
To refactor the `getDish` function and address the long method code flaw, we can break it down into smaller functions with specific responsibilities.
The `renderDish` function is responsible for rendering the HTML structure for a single dish item based on the provided item object.
The `renderRatingStars` function generates the HTML for the rating stars based on the rating value of a dish.
The `getDish` function now uses the `renderDish` function to generate the HTML for each dish item and appends it to the `#dishId` element.

- Conclusion:
Refactoring the code By breaking down the code into smaller functions, we improve code readability, maintainability, and promote code reuse. Each function now has a specific responsibility, making the code easier to understand and modify.